Convert app to read library into Realm database

+++ app/src/main/java/net/sigmabeta/chipbox/ui/playlist/PlaylistFragment.kt
@@ -86,7 +86,7 @@ class PlaylistFragment : BaseFragment(), PlaylistFragmentView, ItemListView<Play
      * ItemListView
      */
 
-    override fun onItemClick(position: Long, clickedViewHolder: PlaylistTrackViewHolder) {
+    override fun onItemClick(position: Int, clickedViewHolder: PlaylistTrackViewHolder) {
         presenter.onItemClick(position)
     }
